添加XML消息  测试群XML消息成功

+++ QQ.Framework/Packets/Send/Message/Send_0x01C0.cs
@@ -40,7 +40,7 @@ namespace QQ.Framework.Packets.Send.Message
         /// <param name="buf">The buf.</param>
         protected override void PutBody()
         {
-            bodyWriter.Write(0x01);
+            bodyWriter.Write((byte)0x01);
             bodyWriter.Write(_toQQ);
         }
     }
